Cleanup svc-assessor-7101v2 docker image (#656)
* [API-1948] conform to gradle standard

* [API-1948] conform to gradle standard

* [API-1234] fix unit tests

* [API-1234] remove requirements

* [API-1234] add medications and recent bp count

* [API-1234] change queue name

* [API-1234] new pdf data

* [API-1234] addd conditions remove medication

* [API-1234] adjust code set

* [API-1234] fix queue test

* fix end to end tests

* remove snomed

* isort

* call new queue and store sufficiency result

* Delete main.py
