+++ app/src/docker/docker-compose.yml
@@ -155,17 +155,6 @@ services:
     networks:
       - internal
 
-  svc-assessor-dc7101v2:
-    image: va/abd_vro-assessclaimdc7101v2:latest
-    <<: *common-security-opt
-    environment:
-      <<: *rabbitmq-placeholder-vars
-      <<: *rabbitmq-client-vars
-    depends_on:
-      rabbitmq-service:
-        condition: service_healthy
-    networks:
-      - internal
   svc-assessor-dc6602:
     image: va/abd_vro-assessclaimdc6602:latest
     <<: *common-security-opt
