+++ build.gradle
@@ -125,7 +125,6 @@ ext {
     ':app',
     ':db-init',
     ':service-python:assessclaimdc7101',
-    ':service-python:assessclaimdc7101v2',
     ':service-python:assessclaimdc6602',
     ':service-python:pdfgenerator',
     ':service-python:featuretoggle',
