+++ service-python/assessclaimdc7101/src/lib/codesets/hypertension_conditions.py
@@ -1,0 +1,6 @@
+conditions = [
+    "I10",
+    "401.0",
+    "401.1",
+    "401.9"
+]
