+++ service-python/assessclaimdc7101/src/lib/settings.py
@@ -1,4 +1,5 @@
 queue_config = {
     "exchange_name": "health-assess-exchange",
     "service_queue_name": "health-assess.7101",
+    "sufficient_queue_name": "health-sufficiency-assess.7101"
 }
