--- service-python/assessclaimdc7101v2/build.gradle
@@ -1,3 +1,0 @@
-plugins {
-  id 'local.python.container-service-conventions'
-}
