--- service-python/assessclaimdc7101v2/src/lib/codesets/hypertension_conditions.py
@@ -1,206 +1,0 @@
-conditions = [
-    "G93.2",
-    "H40.05",
-    "H40.051",
-    "H40.052",
-    "H40.053",
-    "H40.059",
-    "I10",
-    "I12.0",
-    "I12.9",
-    "I15.0",
-    "I15.1",
-    "I15.2",
-    "I15.8",
-    "I27.0",
-    "I27.21",
-    "I27.22",
-    "I27.23",
-    "I27.24",
-    "I27.29",
-    "I87.301",
-    "I87.302",
-    "I87.303",
-    "I87.309",
-    "I87.311",
-    "I87.312",
-    "I87.313",
-    "I87.319",
-    "I87.321",
-    "I87.322",
-    "I87.323",
-    "I87.329",
-    "I87.331",
-    "I87.332",
-    "I87.333",
-    "I87.339",
-    "I87.391",
-    "I87.392",
-    "I87.393",
-    "I87.399",
-    "K76.6",
-    "1201005",
-    "4210003",
-    "8762007",
-    "10725009",
-    "10964002",
-    "11399002",
-    "19769006",
-    "26078007",
-    "28119000",
-    "31992008",
-    "34742003",
-    "38341003",
-    "38481006",
-    "39018007",
-    "39727004",
-    "46481004",
-    "48146000",
-    "48552006",
-    "56218007",
-    "57684003",
-    "59621000",
-    "59720008",
-    "59720008",
-    "63287004",
-    "65518004",
-    "68267002",
-    "70272006",
-    "70995007",
-    "73410007",
-    "74451002",
-    "83291003",
-    "84094009",
-    "86041002",
-    "87394009",
-    "88223008",
-    "89242004",
-    "89242004",
-    "95343001",
-    "123799005",
-    "194783001",
-    "194785008",
-    "194788005",
-    "198941007",
-    "198949009",
-    "233942004",
-    "233943009",
-    "233944003",
-    "233946001",
-    "233947005",
-    "233948000",
-    "233950008",
-    "234072000",
-    "234073005",
-    "234074004",
-    "234075003",
-    "275516004",
-    "276792008",
-    "276793003",
-    "276794009",
-    "360573001",
-    "360578005",
-    "371125006",
-    "397748008",
-    "401047000",
-    "402832006",
-    "423674003",
-    "427889009",
-    "428575007",
-    "429198000",
-    "429457004",
-    "445237003",
-    "449759005",
-    "461321006",
-    "472757001",
-    "472759003",
-    "472760008",
-    "472761007",
-    "472790001",
-    "697896007",
-    "697897003",
-    "697900005",
-    "697901009",
-    "697902002",
-    "697903007",
-    "697904001",
-    "697905000",
-    "697907008",
-    "697908003",
-    "697910001",
-    "697911002",
-    "697912009",
-    "697913004",
-    "697914005",
-    "697917003",
-    "697918008",
-    "697919000",
-    "697920006",
-    "697921005",
-    "697922003",
-    "697923008",
-    "697925001",
-    "697926000",
-    "697927009",
-    "697927009",
-    "697928004",
-    "698838008",
-    "698839000",
-    "703355003",
-    "704667004",
-    "704667004",
-    "707413005",
-    "709666005",
-    "712487000",
-    "712832005",
-    "717824007",
-    "720568003",
-    "722867009",
-    "766937004",
-    "776416004",
-    "840306007",
-    "840307003",
-    "840308008",
-    "870324000",
-    "871642009",
-    "879814007",
-    "879815008",
-    "879816009",
-    "1010625007",
-    "1010627004",
-    "1186852004",
-    "1197664001",
-    "1208839002",
-    "1208845005",
-    "541000000000",
-    "71421000000000",
-    "71701000000000",
-    "104931000000000",
-    "108701000000000",
-    "117681000000000",
-    "127991000000000",
-    "128001000000000",
-    "129151000000000",
-    "129161000000000",
-    "129171000000000",
-    "129181000000000",
-    "140101000000000",
-    "140111000000000",
-    "140121000000000",
-    "140131000000000",
-    "153811000000000",
-    "153831000000000",
-    "153871000000000",
-    "153891000000000",
-    "284961000000000",
-    "284971000000000",
-    "284981000000000",
-    "284991000000000",
-    "285001000000000",
-    "285011000000000",
-    "340441000000000",
-    "345771000000000",
-    "367821000000000",
-    "1078300000000000",
-    "16229400000000000"
-]
