--- service-python/assessclaimdc7101v2/src/lib/settings.py
@@ -1,4 +1,0 @@
-queue_config = {
-    "exchange_name": "health-assess-exchange",
-    "service_queue_name": "health-assess.7101v2",
-}
