--- service-python/assessclaimdc7101v2/src/requirements.txt
@@ -1,3 +1,0 @@
-cerberus==1.3.4;
-pika==1.3.1
-python-dateutil==2.8.2;
