+++ service-python/requirements.txt
@@ -1,5 +1,4 @@
 -r assessclaimdc6602/src/requirements.txt
 -r assessclaimdc7101/src/requirements.txt
--r assessclaimdc7101v2/src/requirements.txt
 -r pdfgenerator/src/requirements.txt
 -r featuretoggle/src/requirements.txt
