+++ service-python/tests/assessclaim/test_queues.py
@@ -4,20 +4,15 @@ from unittest.mock import Mock, patch
 
 import pytest
 
-from assessclaimdc7101v2.src.lib import queues as q7101v2
-from assessclaimdc7101v2.src.lib.main import assess_hypertension as main7101v2
 from assessclaimdc6602.src.lib import queues as q6602
 from assessclaimdc6602.src.lib.main import assess_asthma as main6602
 from assessclaimdc6602v2.src.lib import queues as q6602v2
 from assessclaimdc6602v2.src.lib.main import assess_asthma as main6602v2
 from assessclaimdc7101.src.lib import queues as q7101
-from assessclaimdc7101.src.lib.main import assess_hypertension as main7101
 
 
 @pytest.mark.parametrize(
-    "queue, service_queue_name", [
-        (q6602v2, "6602v2"), (q6602, "6602"), (q7101, "7101"), (q7101v2, "7101v2")
-    ]
+    "queue, service_queue_name", [(q6602, "6602"), (q7101, "7101")]
 )
 def test_queue_setup(queue, service_queue_name, caplog):
     queue_name = f"health-assess.{service_queue_name}"
@@ -48,9 +43,7 @@ def test_queue_setup(queue, service_queue_name, caplog):
     "queue, diagnosticCode, body, main",
     [
         (q6602, "6602", {"evidence": "some medical data body"}, main6602),
-        (q7101, "7101", {"evidence": "some medical data body"}, main7101),
         (q6602v2, "6602v2", {"evidence": "some medical data body"}, main6602v2),
-        (q7101v2, "7101v2", {"evidence": "some medical data body"}, main7101v2),
     ],
 )
 def test_on_request_callback(queue, diagnosticCode, body, main, caplog):
