+++ settings.gradle
@@ -87,7 +87,6 @@ include ':shared:domain'
 include ':shared:lib-camel-connector'
 include ':service-python'
 include ':service-python:assessclaimdc7101'
-include ':service-python:assessclaimdc7101v2'
 include ':service-python:assessclaimdc6602'
 include ':service-python:pdfgenerator'
 include ':service-python:featuretoggle'
